requirements.txt
# Game Analysis Dependencies
gitpython>=3.1.0
requests>=2.31.0
beautifulsoup4>=4.12.0
pillow>=10.0.0
opencv-python>=4.8.0

# AI/ML Dependencies
torch>=2.0.0
transformers>=4.30.0
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0

# Game Development
pygame>=2.5.0
panda3d>=1.10.0

# Web Framework
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0

# Development Tools
pytest>=7.4.0
black>=23.0.0
isort>=5.12.0